Feat(#12):폴더구조 컨밴션대로 수정
