+++ src/app/mypage/edit/page.tsx
@@ -1,0 +1,18 @@
+import { ProfileEditForm, ProfileImageEdit } from '@/components/mypage/(edit)'
+
+export default function page() {
+  return (
+    <div>
+      <h1 className="font-semibold text-[32px]">내 프로필</h1>
+      <div>
+        <div className="flex justify-center mb-11.5">
+          <ProfileImageEdit />
+        </div>
+      </div>
+
+      <div className="flex justify-center">
+        <ProfileEditForm />
+      </div>
+    </div>
+  )
+}
