+++ src/components/mypage/(bookmarks)/index.ts
@@ -1,0 +1,2 @@
+export { default as BookmarkCard } from './BookmarkCard/BookmarkCard'
+export { default as DropDown } from './DropDown/DropDown'
