+++ src/components/mypage/(edit)/PreferenceInfo/PreferenceInfo.tsx
