+++ src/components/mypage/(edit)/index.ts
@@ -1,0 +1,9 @@
+export { default as BasicInfo } from './BasicInfo/BasicInfo'
+export { default as BioField } from './BioField/BioField'
+export { default as BirthdayAndMbti } from './BirthdayAndMbti/BirthdayAndMbti'
+export { default as BirthdaySelect } from './BirthdaySelect/BirthdaySelect'
+export { default as FormActionBtn } from './FormActionBtn/FormActionBtn'
+export { default as GenderField } from './GenderField/GenderField'
+export { default as PreferenceInfo } from './PreferenceInfo/PreferenceInfo'
+export { default as ProfileEditForm } from './ProfileEditForm/ProfileEditForm'
+export { default as ProfileImageEdit } from './ProfileImageEdit/ProfileImageEdit'
