+++ src/components/mypage/(posts)/index.ts
@@ -1,0 +1,1 @@
+export { default as PostsCard } from './PostsCard/PostsCard'
