+++ src/components/mypage/(sent)/index.ts
@@ -1,0 +1,1 @@
+export { default as SentCard } from './SentCard/SentCard'
